[lake] Unify lake related package name to lake
