+++ fluss-flink/fluss-flink-common/src/test/java/com/alibaba/fluss/flink/tiering/committer/TestingCommittableSerializer.java
@@ -16,7 +16,7 @@
 
 package com.alibaba.fluss.flink.tiering.committer;
 
-import com.alibaba.fluss.lakehouse.serializer.SimpleVersionedSerializer;
+import com.alibaba.fluss.lake.serializer.SimpleVersionedSerializer;
 
 import org.apache.flink.core.memory.DataInputDeserializer;
 import org.apache.flink.core.memory.DataOutputSerializer;
