+++ fluss-test-coverage/pom.xml
@@ -322,7 +322,7 @@
                                         <exclude>com.alibaba.fluss.tools.ci.*</exclude>
                                         <!-- exclude for fluss-dist -->
                                         <exclude>com.alibaba.fluss.dist.DummyClass</exclude>
-                                        <exclude>com.alibaba.fluss.lakehouse.batch.ArrowRecordBatch</exclude>
+                                        <exclude>com.alibaba.fluss.lake.batch.ArrowRecordBatch</exclude>
                                         <exclude>com.alibaba.fluss.lake.paimon.FlussDataTypeToPaimonDataType</exclude>
                                         <!-- start exclude for flink tiering service -->
                                         <exclude>com.alibaba.fluss.flink.tiering.source.TieringSourceOptions</exclude>
